+++ lib/src/socket_manager.dart
@@ -1,12 +1,12 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:typed_data';
+import 'package:flutter/foundation.dart';
 
 class SocketManager {
   RawDatagramSocket? _socket;
   final _audioStreamController = StreamController<Uint8List>.broadcast();
   final _statusController = StreamController<String>.broadcast();
-  
+
   Stream<Uint8List> get audioStream => _audioStreamController.stream;
   Stream<String> get statusStream => _statusController.stream;
 
@@ -26,10 +26,10 @@ class SocketManager {
   void sendAudio(Uint8List data, String targetIp, int targetPort) {
     if (_socket != null && targetIp.isNotEmpty) {
       try {
-        print('Sending ${data.length} bytes to $targetIp:$targetPort');
+        debugPrint('Sending ${data.length} bytes to $targetIp:$targetPort');
         _socket!.send(data, InternetAddress(targetIp), targetPort);
       } catch (e) {
-        print("Error sending audio to $targetIp: $e");
+        debugPrint("Error sending audio to $targetIp: $e");
         _statusController.add("Error sending audio: $e");
       }
     }